pyproject.toml
[project]
name = "augur"
version = "0.1.0"
description = "A council of legendary investors, summoned on demand. Give it a ticker; read the omens."
requires-python = ">=3.11"
dependencies = [
    "openai>=1.40.0",
    "httpx>=0.27",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.7",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[project.scripts]
augur = "augur.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/augur"]

[tool.ruff]
line-length = 100
target-version = "py311"
